requirements.txt
pyTermTk
appdirs
pyyaml
pygments
